requirements.txt
requests>=2.18.4
https://github.com/LonamiWebs/Telethon/archive/master.zip
ujson
redis
pymongo
dnspython # needed to connect to cloud MongoDB instances
cryptography
#pillow
aiogram
aioredis
aiocron
git+https://github.com/casics/nostril
flask
flask_bootstrap
pysftp
sentry_sdk
Pillow
